fix: Customer model changed ulong to int for id

+++ api/Models/Customer.cs
@@ -16,7 +16,7 @@ namespace api.Models
     public class Customer
     {
         [Key]
-        public ulong CustomerId { get; set; }
+        public int CustomerId { get; set; }
 
         [Required]
         [Column(TypeName = "nvarchar(100)")]
